[shell update]: 细节优化

+++ shell/src/autosl.sh
@@ -1,22 +1,18 @@
 #!/bin/bash
 
 set -e
+cmd_path="$(dirname $(readlink -f $0))"
+src_path="$cmd_path"
+bin_path="$cmd_path/../bin"
 
-src_path="/home/cc/github.com/vydVolk/linux-learning/shell/src"
-bin_path="/home/cc/github.com/vydVolk/linux-learning/shell/bin"
-
-lib_path="$(dirname $(readlink -f $0))/lib"
+lib_path="$cmd_path/lib"
 . "$lib_path/stdout.sh"
 
 for sh in $(ls $src_path/*.sh)
 do
   file_name=${sh%.*}
   file_name=${file_name##*/}
-  if [ -e "$bin_path/$file_name" ]
-  then
-    rm -rf "$bin_path/$file_name"
-  fi
+  [[ -e "$bin_path/$file_name" ]] && rm -rf "$bin_path/$file_name"
 
   ln -s "$sh" "$bin_path/$file_name" && success "$file_name - completed!"
 done
-
